% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{bin_permute_treatment}
\alias{bin_permute_treatment}
\title{Bin-Based Permutation for DR Tests}
\usage{
bin_permute_treatment(treatment, propensity_scores, n_bins = 10L)
}
\arguments{
\item{treatment}{Binary treatment vector.}

\item{propensity_scores}{Propensity score vector.}

\item{n_bins}{Number of bins.}
}
\value{
Permuted treatment vector.
}
\description{
Permutes treatment labels within propensity score bins.
}
\keyword{internal}
