% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel-matrices.R
\name{centre_kernel_matrix}
\alias{centre_kernel_matrix}
\title{Centre a Kernel Matrix}
\usage{
centre_kernel_matrix(K)
}
\arguments{
\item{K}{Square numeric matrix.}
}
\value{
Centred kernel matrix.
}
\description{
Centres a kernel matrix in feature space (double centring).
}
\keyword{internal}
