% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{cluster_observations}
\alias{cluster_observations}
\title{Simple K-Means Clustering for Permutation Groups}
\usage{
cluster_observations(weights, z, n_clusters = "auto")
}
\arguments{
\item{weights}{Weight vector (density ratios or propensity scores).}

\item{z}{Confounder matrix.}

\item{n_clusters}{Number of clusters. If \code{"auto"}, selects by
silhouette score (2 to 10 clusters).}
}
\value{
Integer vector of cluster assignments.
}
\description{
Clusters observations based on conditional density embeddings
using standard k-means on the density ratio weight space.
}
\keyword{internal}
