% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{cross_fit_propensity}
\alias{cross_fit_propensity}
\title{Cross-Fit Propensity Scores}
\usage{
cross_fit_propensity(treatment, covariates, method, n_folds)
}
\arguments{
\item{treatment}{Binary vector.}

\item{covariates}{Numeric matrix.}

\item{method}{Classification method.}

\item{n_folds}{Number of folds.}
}
\value{
Vector of out-of-sample propensity scores.
}
\description{
Cross-Fit Propensity Scores
}
\keyword{internal}
