% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hsic_stat_cpp}
\alias{hsic_stat_cpp}
\title{Compute the HSIC statistic (unweighted)}
\usage{
hsic_stat_cpp(Kx, Ky)
}
\arguments{
\item{Kx}{n x n kernel matrix for X.}

\item{Ky}{n x n kernel matrix for Y.}
}
\value{
Scalar HSIC value.
}
\description{
Biased HSIC estimator: (1/n^2) * trace(K_x_c \%*\% K_y_c)
where K_c = H K H with H = I - 1/n.
}
\keyword{internal}
