% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{linear_kernel_matrix_cpp}
\alias{linear_kernel_matrix_cpp}
\title{Linear kernel matrix}
\usage{
linear_kernel_matrix_cpp(x, y)
}
\arguments{
\item{x}{Numeric matrix (n x d).}

\item{y}{Numeric matrix (m x d).}
}
\value{
n x m kernel matrix.
}
\description{
K(x, y) = x \%*\% t(y)
}
\keyword{internal}
