% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mmd2_biased_cpp}
\alias{mmd2_biased_cpp}
\title{Compute the biased MMD^2 statistic}
\usage{
mmd2_biased_cpp(Kxx, Kyy, Kxy)
}
\arguments{
\item{Kxx}{n x n kernel matrix for sample X.}

\item{Kyy}{m x m kernel matrix for sample Y.}

\item{Kxy}{n x m kernel matrix between X and Y.}
}
\value{
Scalar biased MMD^2 value.
}
\description{
Compute the biased MMD^2 statistic
}
\keyword{internal}
