% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mmd2_unbiased_cpp}
\alias{mmd2_unbiased_cpp}
\title{Compute the unbiased MMD^2 statistic}
\usage{
mmd2_unbiased_cpp(Kxx, Kyy, Kxy)
}
\arguments{
\item{Kxx}{n x n kernel matrix for sample X.}

\item{Kyy}{m x m kernel matrix for sample Y.}

\item{Kxy}{n x m kernel matrix between X and Y.}
}
\value{
Scalar unbiased MMD^2 value.
}
\description{
Unbiased estimator of \eqn{MMD^2}:
\deqn{\frac{1}{n(n-1)} \sum_{i \neq j} K_{xx}(i,j) + \frac{1}{m(m-1)} \sum_{i \neq j} K_{yy}(i,j) - \frac{2}{nm} \sum_{i,j} K_{xy}(i,j)}
}
\keyword{internal}
