% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permutation_mmd_cpp}
\alias{permutation_mmd_cpp}
\title{Permutation MMD: compute MMD^2 for many permutations of pooled data}
\usage{
permutation_mmd_cpp(K_pool, n, m, n_perm)
}
\arguments{
\item{K_pool}{(n+m) x (n+m) kernel matrix.}

\item{n}{Size of first sample.}

\item{m}{Size of second sample.}

\item{n_perm}{Number of permutations.}
}
\value{
Vector of n_perm MMD^2 values under permutation.
}
\description{
Given the full (n+m) x (n+m) kernel matrix of the pooled sample,
permute the assignment into two groups and compute MMD^2.
}
\keyword{internal}
