% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{polynomial_kernel_matrix_cpp}
\alias{polynomial_kernel_matrix_cpp}
\title{Polynomial kernel matrix}
\usage{
polynomial_kernel_matrix_cpp(x, y, degree, offset)
}
\arguments{
\item{x}{Numeric matrix (n x d).}

\item{y}{Numeric matrix (m x d).}

\item{degree}{Integer degree.}

\item{offset}{Scalar offset.}
}
\value{
n x m kernel matrix.
}
\description{
K(x, y) = (x \%*\% t(y) + offset)^degree
}
\keyword{internal}
