% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbf_kernel_matrix_cpp}
\alias{rbf_kernel_matrix_cpp}
\title{RBF (Gaussian) kernel matrix}
\usage{
rbf_kernel_matrix_cpp(x, y, bandwidth)
}
\arguments{
\item{x}{Numeric matrix (n x d).}

\item{y}{Numeric matrix (m x d).}

\item{bandwidth}{Positive scalar.}
}
\value{
n x m kernel matrix.
}
\description{
K(x, y) = exp(-||x - y||^2 / (2 * bandwidth^2))
}
\keyword{internal}
