% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stratified_permute_cpp}
\alias{stratified_permute_cpp}
\title{Stratified permutation of indices within bins}
\usage{
stratified_permute_cpp(bins, n_bins)
}
\arguments{
\item{bins}{Integer vector of bin assignments (1-indexed).}

\item{n_bins}{Number of bins.}
}
\value{
Permuted index vector (0-indexed for C++).
}
\description{
Given bin assignments, permute indices within each bin.
}
\keyword{internal}
