% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weighted_hsic_stat_cpp}
\alias{weighted_hsic_stat_cpp}
\title{Compute the weighted HSIC statistic (for bd-HSIC)}
\usage{
weighted_hsic_stat_cpp(Kx, Ky, w)
}
\arguments{
\item{Kx}{n x n kernel matrix for X.}

\item{Ky}{n x n kernel matrix for Y.}

\item{w}{Weight vector of length n.}
}
\value{
Scalar weighted HSIC value.
}
\description{
Weighted version: \eqn{\sum_{i,j} w_i \, w_j \, K_{xc,ij} \, K_{yc,ij}}
}
\keyword{internal}
