% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-drdate.R
\name{compute_dr_date_stat}
\alias{compute_dr_date_stat}
\title{Compute DR-DATE Test Statistic}
\usage{
compute_dr_date_stat(Ky, treatment, e_hat, mu_hat_1, mu_hat_0)
}
\arguments{
\item{Ky}{n x n outcome kernel matrix.}

\item{treatment}{Binary treatment vector.}

\item{e_hat}{Propensity scores.}

\item{mu_hat_1}{CME predictions for treated (n x n matrix).}

\item{mu_hat_0}{CME predictions for control (n x n matrix).}
}
\value{
Scalar test statistic.
}
\description{
Compute DR-DATE Test Statistic
}
\keyword{internal}
