% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-drdett.R
\name{compute_dr_dett_stat}
\alias{compute_dr_dett_stat}
\title{Compute DR-DETT Test Statistic}
\usage{
compute_dr_dett_stat(Ky, treatment, e_hat)
}
\arguments{
\item{Ky}{n x n outcome kernel matrix.}

\item{treatment}{Binary treatment vector.}

\item{e_hat}{Propensity scores.}
}
\value{
Scalar test statistic.
}
\description{
Compute DR-DETT Test Statistic
}
\keyword{internal}
