% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{effective_sample_size}
\alias{effective_sample_size}
\title{Compute Effective Sample Size}
\usage{
effective_sample_size(w)
}
\arguments{
\item{w}{Numeric vector of weights.}
}
\value{
Scalar effective sample size.
}
\description{
Computes ESS from importance weights: ESS = (sum(w))^2 / sum(w^2).
}
\examples{
w <- runif(100, 0.5, 2)
effective_sample_size(w)

}
\seealso{
Other density ratio and propensity: 
\code{\link{assess_overlap}()},
\code{\link{estimate_density_ratio}()},
\code{\link{estimate_propensity}()},
\code{\link{fit_density_ratio}()},
\code{\link{plot_weights}()},
\code{\link{predict_density_ratio}()}
}
\concept{density ratio and propensity}
