% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_weights}
\alias{plot_weights}
\title{Plot Weight Diagnostics}
\usage{
plot_weights(weights, main = "Weight Distribution")
}
\arguments{
\item{weights}{Numeric vector of importance weights.}

\item{main}{Title. Default is "Weight Distribution".}
}
\value{
Invisibly returns \code{weights}.
}
\description{
Plots the distribution of importance weights with effective sample
size annotation.
}
\examples{
set.seed(1L)
weights <- rgamma(200L, shape = 2, rate = 2)
plot_weights(weights)

}
\seealso{
Other density ratio and propensity: 
\code{\link{assess_overlap}()},
\code{\link{effective_sample_size}()},
\code{\link{estimate_density_ratio}()},
\code{\link{estimate_propensity}()},
\code{\link{fit_density_ratio}()},
\code{\link{predict_density_ratio}()}
}
\concept{density ratio and propensity}
